How to use Pillow (PIL: Python Imaging Library)
Created functions to extract image data, to apply filter, rotate, and
grayscale effects, save images, and to draw a new image with free fonts.

This is the introduction to the Pillow module.
Image manipulation demands the extraction of data about the image, and
the application of tools to change the image object into desired states.
To this end, several functions were created to load, store, know, and
manipulate image.