Crop a part of the image with Python, Pillow (trimming)
We cropped the image custom, center, and max square.

Cropping means to slice out the image across some dimensions, like
cutting out only the part we want from an image. We have cropped the
image along custom dimensions, and created functions to crop the image
along the center, and towards the max square.